user can now use cart without logging in

+++ app/Http/Controllers/BasketController.php
@@ -12,13 +12,27 @@ use Illuminate\Support\Facades\Auth;
 
 class BasketController extends Controller
 {
-    public function index() {
-        $user_id = Auth::id();
-        $basket = cart::where('user_id', $user_id)->get();
+    public function index(Request $request) {
         $books = collect();
+        if (Auth::check()) {
+            $user_id = Auth::id();
+            $basket = cart::where('user_id', $user_id)->get();
+            
+        } else {
+            if ($request->session()->has('books')) {
+                $basket = $request->session()->get('books');
+            } else {
+                $basket = collect();
+            }
+        }
         //$amounts = array();
+
         foreach ($basket as $elem) {
-            $books->push(Book::where('id', $elem['book_id'])->get());
+            if (Auth::check()) {
+                $books->push(Book::where('id', $elem['book_id'])->get());
+            } else {
+                $books->push(Book::where('id', $elem)->get());
+            }
             //array_push($amounts, $elem['Amount']);     
         }
         return view('/basket', [
@@ -27,27 +41,42 @@ class BasketController extends Controller
         ]);
     }
 
-    public function store($book_id) {
-        $user_id = Auth::id();
-
-        /*$Amount = cart::where('book_id', $book_id)->where('user_id', $user_id)->get('Amount');
-        if ($Amount != "[]") {
-            $basket = cart::where('book_id', $book_id)->where('user_id', $user_id)->get();
-            $basket[0]->Amount = $Amount[0]->Amount + 1;
-            $basket[0]->update();
-        } else {*/
-            $basket = new cart;
-            $basket->user_id = $user_id;
-            $basket->book_id = $book_id;
-            //$basket->Amount = 1;
-            $basket->save();
-        //}
+    public function store($book_id, Request $request) {
+        if (Auth::check()) {
+            $user_id = Auth::id();
+            /*$Amount = cart::where('book_id', $book_id)->where('user_id', $user_id)->get('Amount');
+            if ($Amount != "[]") {
+                $basket = cart::where('book_id', $book_id)->where('user_id', $user_id)->get();
+                $basket[0]->Amount = $Amount[0]->Amount + 1;
+                $basket[0]->update();
+            } else {*/
+                $basket = new cart;
+                $basket->user_id = $user_id;
+                $basket->book_id = $book_id;
+                //$basket->Amount = 1;
+                $basket->save();
+            //}
+        } else {
+            if ($request->session()->has('books')) {
+                $request->session()->push('books', $book_id);
+            } else {
+                $request->session()->put('books', collect());
+            }
+        }
 
         return redirect('basket');
     }
-    public function destroy($book_id) {
-        $user_id = Auth::id();
-        $item = cart::where('book_id', $book_id)->where('user_id', $user_id)->delete();;
+    public function destroy($book_id, Request $request) {
+        if (Auth::check()){
+            $user_id = Auth::id();
+            $item = cart::where('book_id', $book_id)->where('user_id', $user_id)->delete();;
+        } else {
+            $basket = $request->session()->get('books');
+            $key = $basket->search($book_id);
+            $basket->forget($key);
+            $request->session()->put('books', $basket);
+        }
         return redirect('basket');
     }
 }
+
